// src/types/flags.ts
export enum SponsorshipTransferFlags {
  tfSponsorshipEnd = 0x00010000,
  tfSponsorshipCreate = 0x00020000,
  tfSponsorshipReassign = 0x00040000,
}